readyup|refactor: Swap the initCommand tests onto their conventional tails
The two `initCommand` suites carried tails that inverted the idiom the `list/` and `verify/` siblings set: the file named `.unit.` mocked every collaborator, while the file exercising the real command against a temp directory carried no tier at all.

The mocked suite is now `initCommand.wiring.unit.test.ts`, freeing the plain `.unit.` tail for the suite that drives the command itself.
